pyproject.toml

[tool.poetry]
name = "teaseme-backend"
version = "0.1.0"
description = "TeaseMe chat backend (FastAPI + Parlant)"
authors = ["Kako"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
asyncpg = "*"
psycopg-binary = "*"
pgvector = "*"
python-jose = {extras = ["cryptography"], version = "*"}
pydantic-settings = "*"
langchain = "^0.3.26"
openai = "^1.95.1"
langchain-community = "^0.3.27"
langchain-openai = "^0.3.27"
langchain-core = "^0.3.68"
greenlet = "^3.2.3"
psycopg2-binary = "^2.9.10"
python-multipart = "^0.0.20"
pywebpush = "^2.0.3"
cryptography = "^45.0.5"
boto3 = "^1.39.15"
bcrypt = "<4.1.0"
passlib = "<1.8.0"
aiosmtplib = "^4.0.1"
jinja2 = "^3.1.6"
httpx = {extras = ["http2"], version = "^0.28.1"}
redis = "^6.4.0"

[tool.poetry.group.dev.dependencies]
alembic = "*"
ruff = "*"
black = "*"
mypy = "*"
pytest = "*"
python-dotenv = "^1.1.1"
